cloud/aws: move maintenance calls to separate file
